File Server Details Pop up
Reviewers       :       prism ui
Tickets resolved        :     ENG-304378
Tickets related :       None
Tests run       :       Manual
Target release  :       5.17.1
Breaks rolling upgrade  : No

Link to the ticket:
https://jira.nutanix.com/browse/ENG-304378

Description:
Rename the "Done" to "Close"
Remove the manage button
Add the link to the Manage File Server the top of the table. Use a blue link with the icon similar to the one on the File server table.